+++ demisto_sdk/__main__.py
@@ -365,11 +365,7 @@ def unify(**kwargs):
     is_flag=True)
 @click.option(
     "--allow-skipped",
-    help="Don't fail on skipped integrations",
-    is_flag=True)
-@click.option(
-    "--allow-all-skipped-playbooks",
-    help="Don't fail when all test playbooks are skipped",
+    help="Don't fail on skipped integrations or when all test playbooks are skipped",
     is_flag=True)
 @pass_config
 def validate(config, **kwargs):
@@ -410,8 +406,7 @@ def validate(config, **kwargs):
             debug_git=kwargs.get('debug_git'),
             include_untracked=kwargs.get('include_untracked'),
             quite_bc=kwargs.get('quite_bc_validation'),
-            check_is_unskipped=not kwargs.get('allow_skipped', False),
-            check_unskipped_playbooks=not kwargs.get('allow_all_skipped_playbooks', False)
+            check_is_unskipped=not kwargs.get('allow_skipped', False)
         )
         return validator.run_validation()
     except (git.InvalidGitRepositoryError, git.NoSuchPathError, FileNotFoundError) as e: