+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -432,7 +432,7 @@ class TestIntegrationValidation:
         pack_integration_path = join(AZURE_FEED_PACK_PATH, "Integrations/FeedAzure/FeedAzure.yml")
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(main, [VALIDATE_CMD, "-i", pack_integration_path, "--no-conf-json",
-                                      "--allow-all-skipped-playbooks"])
+                                      "--allow-skipped"])
 
         assert f"Validating {pack_integration_path} as integration" in result.stdout
         assert "The docker image tag is not the latest numeric tag, please update it" in result.stdout
@@ -455,7 +455,7 @@ class TestIntegrationValidation:
         integration_path = join(TEST_FILES_PATH, 'integration-invalid-no-hidden-params.yml')
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(main, [VALIDATE_CMD, "-i", integration_path, "--no-conf-json",
-                                      "--allow-all-skipped-playbooks"])
+                                      "--allow-skipped"])
         assert result.exit_code == 1
         assert f"Validating {integration_path} as integration" in result.stdout
         assert "can't be hidden. Please remove this field" in result.stdout
@@ -530,7 +530,7 @@ class TestPackValidation:
         mocker.patch.object(IntegrationValidator, 'is_there_separators_in_names', return_value=True)
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(main, [VALIDATE_CMD, "-i", VALID_PACK_PATH, "--no-conf-json",
-                                      "--allow-all-skipped-playbooks"])
+                                      "--allow-skipped"])
         assert f"{VALID_PACK_PATH} unique pack files" in result.stdout
         assert f"Validating pack {VALID_PACK_PATH}" in result.stdout
         assert f"{VALID_PACK_PATH}/Integrations/FeedAzureValid/FeedAzureValid.yml" in result.stdout
@@ -554,7 +554,7 @@ class TestPackValidation:
         mocker.patch.object(BaseValidator, 'check_file_flags', return_value='')
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(main, [VALIDATE_CMD, "-i", AZURE_FEED_PACK_PATH, "--no-conf-json",
-                                      "--allow-all-skipped-playbooks"])
+                                      "--allow-skipped"])
 
         assert f'{AZURE_FEED_PACK_PATH}' in result.output
         assert f'{AZURE_FEED_PACK_PATH}/IncidentFields/incidentfield-city.json' in result.output